+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+- Interactively update any GitHub Actions found in workflows, there is no more default action or custom action
+
 ## [0.0.3] - 2023-06-27
 
 ### Added